+++ src/RepoM.Plugin.Statistics/Ordering/ScoreCalculatorConfig.cs
@@ -1,0 +1,10 @@
+namespace RepoM.Plugin.Statistics.Ordering;
+
+using System.Collections.Generic;
+
+internal class ScoreCalculatorConfig
+{
+    public List<ScoreCalculatorRangeConfig> Ranges { get; set; } = new();
+
+    public int MaxScore { get; set; } = int.MaxValue;
+}
