+++ src/RepoM.Plugin.Statistics/Ordering/ScoreCalculatorRangeConfig.cs
@@ -1,0 +1,12 @@
+namespace RepoM.Plugin.Statistics.Ordering;
+
+using System;
+
+internal class ScoreCalculatorRangeConfig
+{
+    public TimeSpan MaxAge { get; set; }
+
+    public int Score { get; set;}
+
+    public int MaxItems { get; set; } = int.MaxValue;
+}
